staging/pkg/mod/golang.org/x/net@v0.0.0-20201110031124-69a78807bb2b/html/testdata/go/select.dat

#data
<table><math><select><mi><select></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math select>
|         <math mi>
|           <select>
|     <table>
